permission

+++ spatie-app/routes/api.php
@@ -2,69 +2,43 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\TaskController;
+use App\Http\Controllers\UserController;
 use App\Http\Controllers\DepartmentController;
+use App\Http\Controllers\TeamController;
+use App\Http\Controllers\TeamMemberController;
 
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider and all of them will
-| be assigned to the "api" middleware group. Make something great!
-|
-*/
-
-Route::post('/create', [TaskController::class, 'assignTask']);
-Route::post('/show', [TaskController::class, 'show']);
-
-Route::delete('/delete/{id}', [TaskController::class, 'destroy']);
-
-Route::put('/update/{id}', [TaskController::class, 'Update_task']); // for updating complete start
-Route::put('/reassign/{id}', [TaskController::class, 'reassign_task']); // for reassigning the task
-//Route::resource('teams', TeamController::class);
-
-// '/api/team-members' -> this endpoint is for TeamMember
-
-//Route::resource('teammembers', TeamMemberController::class);
-
-
-// Routes for TeamController
-Route::get('/teamindex', [TeamController::class, 'index']);
-
-Route::post('/teamcreate', [TeamController::class, 'store']);
-
-Route::get('/teamshow/{id}', [TeamController::class, 'show']);
-
-Route::put('/teamupdate', [TeamController::class, 'update']);
-
-Route::delete('/teamdestroy/{id}', [TeamController::class, 'destroy']);
-
-// Routes for TeamMemberController
-Route::get('/memberindex', [TeamMemberController::class, 'index']);
-
-Route::post('/membercreate', [TeamMemberController::class, 'store']);
-
-Route::get('/membershow/{id}', [TeamMemberController::class, 'show']);
-
-Route::put('/memberupdate/{id}', [TeamMemberController::class, 'update']);
-
-Route::delete('/memberdestroy/{id}', [TeamMemberController::class, 'destroy']);
-
-
-
+Route::post('/allusers', [UserController::class, 'user'])->middleware('permission:access_user');
 Route::post('/login', [UserController::class, 'login']);
 Route::post('/register', [UserController::class, 'register']);
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::post('add',[DepartmentController::class,'adding']);
+//Routes for DepartmentController
+Route::post('add', [DepartmentController::class, 'adding'])->middleware('permission:create_departments');
+Route::put('update/{id}', [DepartmentController::class, 'updating'])->middleware('permission:update_departments');
+Route::delete('delete/{id}', [DepartmentController::class, 'delete'])->middleware('permission:access_departments');
+Route::get('get', [DepartmentController::class, 'getData'])->middleware('permission:access_departments');
+;
 
-Route::put('update/{id}',[DepartmentController::class,'updating']);
 
-Route::delete('delete/{id}',[DepartmentController::class,'delete']);
+// Routes for TeamController
+Route::get('/teamindex', [TeamController::class, 'index'])->middleware('permission:view_team');
+Route::post('/teamcreate', [TeamController::class, 'store'])->middleware('permission:add_team');
+Route::get('/teamshow/{id}', [TeamController::class, 'show'])->middleware('permission:view_team');
+Route::put('/teamupdate', [TeamController::class, 'update'])->middleware('permission:update_team');
+Route::delete('/teamdestroy/{id}', [TeamController::class, 'destroy'])->middleware('permission:add_team');
 
-Route::get('get',[DepartmentController::class,'getData']);
+
+// Routes for TeamMemberController
+Route::post('/membercreate', [TeamMemberController::class, 'store'])->middleware('permission:add_team');
+Route::post('/membershow', [TeamMemberController::class, 'index'])->middleware('permission:view_team');
+Route::delete('/memberdestroy/{id}', [TeamMemberController::class, 'destroy'])->middleware('permission:add_team');
 
 
+//Routes for TaskController
+Route::post('/taskcreate', [TaskController::class, 'assignTask'])->middleware('permission:assign_task');
+Route::post('/taskshow', [TaskController::class, 'show'])->middleware('permission:view_task');
+Route::delete('/taskdelete/{id}', [TaskController::class, 'destroy'])->middleware('permission:create_task');
+Route::put('/taskupdate/{id}', [TaskController::class, 'Update_task'])->middleware('permission:update_task'); // for updating complete start
+Route::put('/taskreassign/{id}', [TaskController::class, 'reassign_task'])->middleware('permission:assign_task'); // for reassigning the task
